fix: correct usage of size_t in common module

+++ engine/src/common/constants.h
@@ -17,10 +17,10 @@ inline constexpr size_t kVectorMemsize = kVectorDim * sizeof(float);
 inline constexpr float kSimilarityThreshold = 0.85f;
 
 // L0 buffer's slot limits.
-inline constexpr size_t kL0Capacity = 1 << 12;
+inline constexpr uint32_t kL0Capacity = 1 << 12;
 // Total number of slots in the unified `MemoryArena`.
-inline constexpr size_t kTotalSlots = 1 << 19;
+inline constexpr uint32_t kTotalSlots = 1 << 19;
 
 // Maximum lifetime of a `kPending` node in seconds.
 // Nodes remain `kPending` beyond this deadline are treated as stale by the GC.
-inline constexpr uint64_t kPendingLifespan = 30;
+inline constexpr uint32_t kPendingLifespan = 30;
